+++ pkg/server/alert/alert_analyze_test.go
@@ -108,57 +108,6 @@ func TestAnalyzeAlert(t *testing.T) {
 	}
 }
 
-func TestSendSlackNotification(t *testing.T) {
-	httpmock.Activate()
-	defer httpmock.DeactivateAndReset()
-	httpmock.RegisterResponder("POST", "http://hogehoge.com", httpmock.NewStringResponder(200, "mocked"))
-	httpmock.RegisterResponder("POST", "http://fugafuga.com", httpmock.NewErrorResponder(errors.New("Something Wrong")))
-	cases := []struct {
-		name          string
-		notifySetting string
-		alert         *model.Alert
-		project       *project.Project
-		wantErr       bool
-	}{
-		{
-			name:          "OK",
-			notifySetting: `{"webhook_url":"http://hogehoge.com"}`,
-			alert:         &model.Alert{},
-			project:       &project.Project{},
-			wantErr:       false,
-		},
-		{
-			name:          "NG Json.Marshal Error",
-			notifySetting: `{"webhook_url":http://hogehoge.com"}`,
-			alert:         &model.Alert{},
-			project:       &project.Project{},
-			wantErr:       true,
-		},
-		{
-			name:          "Warn webhook_url not set",
-			notifySetting: `{}`,
-			alert:         &model.Alert{},
-			project:       &project.Project{},
-			wantErr:       false,
-		},
-		{
-			name:          "HTTP Error",
-			notifySetting: `{"webhook_url":"http://fugafuga.com"}`,
-			alert:         &model.Alert{},
-			project:       &project.Project{},
-			wantErr:       true,
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			got := sendSlackNotification(context.Background(), "unused", c.notifySetting, c.alert, c.project, &[]model.AlertRule{})
-			if (got != nil && !c.wantErr) || (got == nil && c.wantErr) {
-				t.Fatalf("Unexpected error: %+v", got)
-			}
-		})
-	}
-}
-
 func TestNotificationAlert(t *testing.T) {
 	httpmock.Activate()
 	defer httpmock.DeactivateAndReset()
@@ -545,60 +494,6 @@ func TestRegistAlertByAnalyze(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, c.want) {
 				t.Fatalf("Unexpected response: want=%+v, got=%+v", c.want, got)
-			}
-		})
-	}
-}
-
-func TestGenerateRuleList(t *testing.T) {
-	cases := []struct {
-		name  string
-		input *[]model.AlertRule
-		want  string
-	}{
-		{
-			name: "1 line",
-			input: &[]model.AlertRule{
-				{AlertRuleID: 1, Name: "aaa"},
-			},
-			want: "- aaa",
-		},
-		{
-			name: "Multi lines",
-			input: &[]model.AlertRule{
-				{AlertRuleID: 1, Name: "aaa"},
-				{AlertRuleID: 2, Name: "bbb"},
-				{AlertRuleID: 3, Name: "ccc"},
-			},
-			want: "- aaa\n- bbb\n- ccc",
-		},
-		{
-			name:  "Nil input",
-			input: nil,
-			want:  "",
-		},
-		{
-			name: "Too many rules",
-			input: &[]model.AlertRule{
-				{AlertRuleID: 1, Name: "aaa"},
-				{AlertRuleID: 2, Name: "bbb"},
-				{AlertRuleID: 3, Name: "ccc"},
-				{AlertRuleID: 4, Name: "ddd"},
-				{AlertRuleID: 5, Name: "eee"},
-				{AlertRuleID: 6, Name: "fff"},
-				{AlertRuleID: 7, Name: "ggg"},
-				{AlertRuleID: 8, Name: "hhh"},
-				{AlertRuleID: 9, Name: "iii"},
-				{AlertRuleID: 10, Name: "jjj"},
-			},
-			want: "- aaa\n- bbb\n- ccc\n- ddd\n- eee\n- ...",
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			got := generateRuleList(c.input)
-			if !reflect.DeepEqual(got, c.want) {
-				t.Fatalf("Unexpected result: want=%+v, got=%+v", c.want, got)
 			}
 		})
 	}
